requirements.txt
openai-whisper>=20231117
anthropic>=0.39.0
python-dotenv>=1.0.0
pydub>=0.25.1
